+++ pioreactor/automations/dosing/pid_morbidostat.py
@@ -11,14 +11,12 @@ from pioreactor.utils import local_persistant_storage
 from pioreactor.utils.streaming_calculations import PID
 
 
-VIAL_VOLUME = float(config["bioreactor"]["volume_ml"])
-
-
 class PIDMorbidostat(DosingAutomationJob):
     """
     As defined in Zhong 2020
     """
 
+    VIAL_VOLUME = float(config["bioreactor"]["volume_ml"])
     automation_name = "pid_morbidostat"
     published_settings = {
         "volume": {"datatype": "float", "settable": True, "unit": "mL"},
@@ -66,7 +64,7 @@ class PIDMorbidostat(DosingAutomationJob):
             )
 
         assert isinstance(self.duration, float)
-        self.volume = round(self.target_growth_rate * VIAL_VOLUME * (self.duration / 60), 4)
+        self.volume = round(self.target_growth_rate * self.VIAL_VOLUME * (self.duration / 60), 4)
 
     def execute(self) -> events.AutomationEvent:
         if self.latest_normalized_od <= self.min_od: